Refine download_once to download cookbooks that aren't already there
That is to say in the default case we wipe the whole directory.  Then
we check to see if the subdirectory exists before downloading a
cookbook; if it does exist, then we should not re-download it.  This
should handle both the normal and download_once paths with the same
logic.