LoginPolicy update: public keys management changed
Root's authorized_keys file is not updated anymore. All users'
public keys are written to RightScale user's authorized_keys with
thunking. The difference is that superusers are added to additional
group "rightscale" which has NOPASSWD option for sudo.
Users are able to login with root@hostname only if they have default
public key.
Exclusive option for LoginPolicy is obsolete now.