acu178341 Revert some parts of azure/cloudstack metadata
Polling for the metadata provider in the root of the plugins was getting
evaluated at load time for the plugins for all clouds. For windows this
included 20 minute waits waiting for it to resolve the DHCP server.
Refactored to only poll for the provider if we're on the right cloud.